UI has had some basic changes to the way the buttons and background look.
Logic has been added to the buttons. They now call the Randomise function, to generate a random number based on the value of the button. This has been set to never produce a 0 as a dice wouldnt.

Possible improvements would be to change this to an MVVM data binding style but currently the label is being edited crudely.